+++ CHANGELOG.md
@@ -11,6 +11,10 @@ Changelog for paco.models
 - Added ``IRDSPostgresql`` with RDS for Postgresql support. Added complete list of RDS EngineVerions for
   Mysql and Postgresql to vocabulary.
 
+### Changed
+
+- ``Lambda:code:zipfile`` can now be a path to a local directory.
+
 7.1.0 (2020-04-04)
 ------------------